+++ nexus/pkg/server/connection.go
@@ -2,7 +2,6 @@ package server
 
 import (
 	"bufio"
-	"bytes"
 	"context"
 	"encoding/binary"
 	"net"
@@ -12,17 +11,6 @@ import (
 	"github.com/wandb/wandb/nexus/pkg/service"
 	"google.golang.org/protobuf/proto"
 )
-
-type Header struct {
-	Magic      uint8
-	DataLength uint32
-}
-
-type Tokenizer struct {
-	header       Header
-	headerLength int
-	headerValid  bool
-}
 
 type Connection struct {
 	ctx    context.Context
@@ -50,41 +38,6 @@ func NewConnection(
 		requestChan:  make(chan *service.ServerRequest),
 		respondChan:  make(chan *service.ServerResponse),
 	}
-}
-
-func (x *Tokenizer) split(data []byte, _ bool) (advance int, token []byte, err error) {
-	if x.headerLength == 0 {
-		x.headerLength = binary.Size(x.header)
-	}
-
-	advance = 0
-
-	if !x.headerValid {
-		if len(data) < x.headerLength {
-			return
-		}
-		buf := bytes.NewReader(data)
-		err := binary.Read(buf, binary.LittleEndian, &x.header)
-		if err != nil {
-			log.Error(err)
-			return 0, nil, err
-		}
-		if x.header.Magic != uint8('W') {
-			log.Error("Invalid magic byte in header")
-		}
-		x.headerValid = true
-		advance += x.headerLength
-		data = data[advance:]
-	}
-
-	if len(data) < int(x.header.DataLength) {
-		return
-	}
-
-	advance += int(x.header.DataLength)
-	token = data[:x.header.DataLength]
-	x.headerValid = false
-	return
 }
 
 func (nc *Connection) receive(wg *sync.WaitGroup) {
